Vision Area Auto-Tune in progress
As the limelight will be mounted to the robot in such a manner that the
height difference between the vision tape and the camera is little,
trigonemetric methods for calculating distanec to the target will become
inaccurate. Therefore, forming a "look-up" table was decided on, relying
on linear interpolation to approximate the distance to the target given
target area. Auto-Tuning Code uses a simple routine that leverages the
odometery capabilities of the drivetrain encoders to calibrate this data
table.