Miscellaneous Changes to Subsystems and Vision
Vision auto-tune now gathers appropriate data and publishes it to the
filesystem as a csv using CSVLogger. The strange bug with drivetrain
encoders not zeroing has somehow been resolved but CSVLogger is very
dirty and is in dire need of a clean up and commenting. Overall layout
of the Vision auto-tune is rather messy so clean up is necessary at
some point, however it isn't really a mission critical feature so maybe
hold of on this for a bit. Future steps involve reading the csv into an
arraylist for use as an interpolation table.